+++ app/assets/javascripts/govuk_publishing_components/lib/filter-list.js
@@ -3,13 +3,24 @@
 
   function FilterList (module) {
     this.module = module
-    this.items = this.module.querySelectorAll('[data-filter-item]')
+    this.items = document.querySelectorAll('[data-filter-item]')
     this.labelText = this.module.getAttribute('data-filter-label') || 'Filter list'
+    this.resultsFound = this.items.length
+    this.resultsId = 'filter-results-count'
+
+    window.GOVUK.vars = window.GOVUK.vars || {}
   }
 
   FilterList.prototype.init = function () {
-    if (this.items.length) {
+    if (window.GOVUK.vars.filterStarted) {
+      return
+    }
+
+    if (this.resultsFound) {
       this.appendFilterInput()
+      this.resultsElement = document.getElementById(this.resultsId)
+      this.updateResultsCount()
+      window.GOVUK.vars.filterStarted = true
     }
   }
 
@@ -23,13 +34,19 @@
     formLabel.textContent = this.labelText
 
     const formInput = document.createElement('input')
-    formInput.classList.add('gem-c-input', 'govuk-input')
+    formInput.classList.add('gem-c-input', 'govuk-input', 'govuk-!-margin-bottom-1')
     formInput.id = 'filterInput'
     formInput.name = 'name'
     formInput.spellcheck = false
     formInput.type = 'text'
+    formInput.setAttribute('aria-describedby', this.resultsId)
 
-    form.append(formLabel, formInput)
+    const results = document.createElement('div')
+    results.classList.add('govuk-hint')
+    results.id = this.resultsId
+    results.setAttribute('aria-live', 'polite')
+
+    form.append(formLabel, formInput, results)
 
     this.module.prepend(form)
     const input = form.querySelector('.govuk-input')
@@ -50,6 +67,14 @@
         item.classList.add('govuk-!-display-none')
       }
     }
+    this.updateResultsCount()
+  }
+
+  FilterList.prototype.updateResultsCount = function () {
+    const count = Array.from(this.items).filter((item) => !item.classList.contains('govuk-!-display-none')).length
+    const text = count === 1 ? `${count} result found` : `${count === 0 ? 'No' : count} results found`
+
+    this.resultsElement.innerHTML = count !== this.resultsFound ? text : ''
   }
 
   Modules.FilterList = FilterList