+++ spec/javascripts/govuk_publishing_components/lib/filter-list-spec.js
@@ -4,9 +4,12 @@ describe('The filter list code', function () {
   const el = document.createElement('div')
   let input
   const markup = `
-    <p data-filter-item="a">United Kingdom</p>
-    <p data-filter-item="b">United States</p>
-    <p data-filter-item="c">United Kingdom of Smaller Countries</p>
+    <div id="separate"></div>
+    <div>
+      <p data-filter-item="a">United Kingdom</p>
+      <p data-filter-item="b">United States</p>
+      <p data-filter-item="c">United Kingdom of Smaller Countries</p>
+    </div>
   `
 
   beforeEach(function () {
@@ -14,6 +17,7 @@ describe('The filter list code', function () {
   })
 
   afterEach(function () {
+    delete window.GOVUK.vars.filterStarted
     document.body.removeChild(el)
   })
 
@@ -26,13 +30,30 @@ describe('The filter list code', function () {
   describe('filtering', function () {
     beforeEach(function () {
       el.innerHTML = markup
-      new window.GOVUK.Modules.FilterList(el).init()
+      new window.GOVUK.Modules.FilterList(document.getElementById('separate')).init()
       input = el.querySelector('input')
+    })
+
+    it('only creates one instance of itself', function () {
+      const markup2 = `
+        <p data-filter-item="a">United Kingdom</p>
+        <p data-filter-item="b">United States</p>
+      `
+      const el2 = document.createElement('div')
+      el2.innerHTML = markup2
+      document.body.appendChild(el2)
+      new window.GOVUK.Modules.FilterList(el2).init()
+      expect(document.querySelectorAll('input.gem-c-input')).toHaveSize(1)
+      document.body.removeChild(el2)
     })
 
     it('creates an input if there is something to filter', function () {
       expect(el.querySelectorAll('label.gem-c-label')).toHaveSize(1)
       expect(el.querySelectorAll('input.gem-c-input')).toHaveSize(1)
+      expect(el.querySelector('input.gem-c-input').getAttribute('aria-describedby')).toBe('filter-results-count')
+      expect(el.querySelectorAll('.govuk-hint')).toHaveSize(1)
+      expect(el.querySelector('.govuk-hint').textContent).toBe('')
+      expect(el.querySelector('.govuk-hint').getAttribute('aria-live')).toBe('polite')
     })
 
     it('shows one element based on user input', function () {
@@ -41,6 +62,7 @@ describe('The filter list code', function () {
       expect(el.querySelector('[data-filter-item="a"]')).toHaveClass('govuk-!-display-none')
       expect(el.querySelector('[data-filter-item="b"]')).not.toHaveClass('govuk-!-display-none')
       expect(el.querySelector('[data-filter-item="c"]')).toHaveClass('govuk-!-display-none')
+      expect(el.querySelector('.govuk-hint').textContent).toBe('1 result found')
     })
 
     it('shows multiple elements based on user input', function () {
@@ -49,6 +71,16 @@ describe('The filter list code', function () {
       expect(el.querySelector('[data-filter-item="a"]')).not.toHaveClass('govuk-!-display-none')
       expect(el.querySelector('[data-filter-item="b"]')).toHaveClass('govuk-!-display-none')
       expect(el.querySelector('[data-filter-item="c"]')).not.toHaveClass('govuk-!-display-none')
+      expect(el.querySelector('.govuk-hint').textContent).toBe('2 results found')
+    })
+
+    it('shows no results if the input does not match', function () {
+      input.value = 'france'
+      window.GOVUK.triggerEvent(input, 'input')
+      expect(el.querySelector('[data-filter-item="a"]')).toHaveClass('govuk-!-display-none')
+      expect(el.querySelector('[data-filter-item="b"]')).toHaveClass('govuk-!-display-none')
+      expect(el.querySelector('[data-filter-item="c"]')).toHaveClass('govuk-!-display-none')
+      expect(el.querySelector('.govuk-hint').textContent).toBe('No results found')
     })
 
     it('resets if the input is cleared', function () {
@@ -57,6 +89,7 @@ describe('The filter list code', function () {
       expect(el.querySelector('[data-filter-item="a"]')).not.toHaveClass('govuk-!-display-none')
       expect(el.querySelector('[data-filter-item="b"]')).not.toHaveClass('govuk-!-display-none')
       expect(el.querySelector('[data-filter-item="c"]')).not.toHaveClass('govuk-!-display-none')
+      expect(el.querySelector('.govuk-hint').textContent).toBe('')
     })
   })